pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pydanticai-telegram-bot"
version = "0.1.0"
description = "Simple Telegram bot with PydanticAI, OpenAI-compatible models, and plugin-based tools."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx",
    "pydantic-ai-slim[openai]",
    "python-telegram-bot",
    "starlette",
    "uvicorn",
]

[project.scripts]
telegram-bot = "app.bot.telegram_app:main"
agent-web = "app.web.app:main"
agent-all = "app.run_all:main"
caldav-tool = "app.plugins.caldav.cli:main"
get-time-tool = "app.plugins.get_time.cli:main"
intervals-icu-tool = "app.plugins.intervals_icu.cli:main"
open-meteo-tool = "app.plugins.open_meteo.cli:main"
route-planner-tool = "app.plugins.route_planner.cli:main"

[dependency-groups]
dev = [
    "pytest",
    "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
